create AbstractResourceIntegrationTest to avoid duplicate code #67
