+++ storage/src/test/java/org/zalando/apidiscovery/storage/AbstractResourceIntegrationTest.java
@@ -1,0 +1,45 @@
+package org.zalando.apidiscovery.storage;
+
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.context.embedded.LocalServerPort;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.web.util.UriComponentsBuilder;
+import org.zalando.apidiscovery.storage.api.ApiRepository;
+import org.zalando.apidiscovery.storage.api.ApplicationRepository;
+
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
+
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest(webEnvironment = RANDOM_PORT)
+public abstract class AbstractResourceIntegrationTest {
+
+    @Autowired
+    protected TestRestTemplate restTemplate;
+
+    @Autowired
+    protected ApplicationRepository applicationRepository;
+
+    @Autowired
+    protected ApiRepository apiRepository;
+
+    @LocalServerPort
+    protected int port;
+
+    @Before
+    public void cleanDatabase() {
+        applicationRepository.deleteAll();
+        apiRepository.deleteAll();
+    }
+
+    public UriComponentsBuilder localUriBuilder() {
+        return UriComponentsBuilder.newInstance()
+            .scheme("http")
+            .host("localhost")
+            .port(port);
+    }
+}
