+++ storage/src/test/java/org/zalando/apidiscovery/storage/ApiDefinitionResourceIntegrationTest.java
@@ -1,0 +1,6 @@
+package org.zalando.apidiscovery.storage;
+
+public class ApiDefinitionResourceIntegrationTest {
+
+
+}
