+++ storage/src/test/java/org/zalando/apidiscovery/storage/ApiResourceIntegrationTest.java
@@ -1,22 +1,12 @@
 package org.zalando.apidiscovery.storage;
 
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.embedded.LocalServerPort;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.web.util.UriComponentsBuilder;
 import org.zalando.apidiscovery.storage.api.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.ApiEntity;
 import org.zalando.apidiscovery.storage.api.ApiLifecycleState;
-import org.zalando.apidiscovery.storage.api.ApiRepository;
 import org.zalando.apidiscovery.storage.api.ApplicationEntity;
-import org.zalando.apidiscovery.storage.api.ApplicationRepository;
 
 import static com.jayway.jsonpath.matchers.JsonPathMatchers.hasJsonPath;
 import static com.jayway.jsonpath.matchers.JsonPathMatchers.isJson;
@@ -30,14 +20,11 @@ import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.hasSize;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static org.zalando.apidiscovery.storage.ApiLifecycleManager.ACTIVE;
 import static org.zalando.apidiscovery.storage.ApiLifecycleManager.DECOMMISSIONED;
 import static org.zalando.apidiscovery.storage.ApiLifecycleManager.INACTIVE;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT)
-public class ApiResourceIntegrationTest {
+public class ApiResourceIntegrationTest extends AbstractResourceIntegrationTest {
 
     private final static String TEST_API = "testAPI";
     private final static String ANOTHER_API = "anotherAPI";
@@ -45,25 +32,6 @@ public class ApiResourceIntegrationTest {
     private final static String V2 = "2.0.0";
     private final static String APP1 = "app1";
     private final static String APP2 = "app2";
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Autowired
-    private ApplicationRepository applicationRepository;
-
-    @Autowired
-    private ApiRepository apiRepository;
-
-    @LocalServerPort
-    private int port;
-
-
-    @Before
-    public void cleanDatabase() {
-        applicationRepository.deleteAll();
-        apiRepository.deleteAll();
-    }
 
     @Test
     public void shouldReturnAllApis() throws Exception {
@@ -303,13 +271,6 @@ public class ApiResourceIntegrationTest {
         ResponseEntity<String> responseEntity = restTemplate.getForEntity(
             "/apis/" + TEST_API, String.class);
         assertThat(responseEntity.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
-    }
-
-    private UriComponentsBuilder localUriBuilder() {
-        return UriComponentsBuilder.newInstance()
-            .scheme("http")
-            .host("localhost")
-            .port(port);
     }
 
 }
