+++ storage/src/test/java/org/zalando/apidiscovery/storage/ApiVersionResourceIntegrationTest.java
@@ -1,21 +1,11 @@
 package org.zalando.apidiscovery.storage;
 
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.embedded.LocalServerPort;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.web.util.UriComponentsBuilder;
 import org.zalando.apidiscovery.storage.api.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.ApiEntity;
-import org.zalando.apidiscovery.storage.api.ApiRepository;
 import org.zalando.apidiscovery.storage.api.ApplicationEntity;
-import org.zalando.apidiscovery.storage.api.ApplicationRepository;
 
 import static com.jayway.jsonpath.matchers.JsonPathMatchers.hasJsonPath;
 import static java.util.Arrays.asList;
@@ -24,7 +14,6 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasItems;
 import static org.hamcrest.Matchers.hasSize;
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static org.zalando.apidiscovery.storage.DomainObjectGen.API_NAME;
 import static org.zalando.apidiscovery.storage.DomainObjectGen.API_UI;
 import static org.zalando.apidiscovery.storage.DomainObjectGen.API_URL;
@@ -40,28 +29,7 @@ import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.ACTIVE;
 import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.DECOMMISSIONED;
 import static org.zalando.apidiscovery.storage.api.ApiLifecycleState.INACTIVE;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT)
-public class ApiVersionResourceIntegrationTest {
-
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-    @Autowired
-    private ApplicationRepository applicationRepository;
-
-    @Autowired
-    private ApiRepository apiRepository;
-
-    @LocalServerPort
-    private int port;
-
-
-    @Before
-    public void cleanDatabase() {
-        applicationRepository.deleteAll();
-        apiRepository.deleteAll();
-    }
+public class ApiVersionResourceIntegrationTest extends AbstractResourceIntegrationTest {
 
     @Test
     public void shouldGroupVersionOfGivenApi() throws Exception {
@@ -226,13 +194,5 @@ public class ApiVersionResourceIntegrationTest {
 
         assertThat(responseEntity.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
     }
-
-    private UriComponentsBuilder localUriBuilder() {
-        return UriComponentsBuilder.newInstance()
-            .scheme("http")
-            .host("localhost")
-            .port(port);
-    }
-
 
 }
